\name{Crossprod}
\alias{Crossprod}
\title{Crossproduct of vectors}
\description{
This function performs the cross product of two three-dimensional vectors.
}
\usage{
Crossprod(vec1, vec2)
}
\arguments{
  \item{vec1}{A three-dimensional vector.}
  \item{vec2}{A three-dimensional vector.}
}
\details{
This function performs the cross product of two three-dimensional vectors.
}
\value{
three-dimensional vector.
}
\author{
Jongmin Lee
}
\examples{
Crossprod(c(1, 1, 1), c(5,6,10))
}
